requirements.txt

fastapi
uvicorn
pandas
numpy
faiss-cpu
sentence-transformers
scikit-learn
transformers
requests
beautifulsoup4
python-json-logger
